requirements.txt
openai==2.31.0
pydantic==2.8.2
pandas==2.2.2
python-dotenv==0.21.0
tiktoken==0.8.0
